Removed remaining submodules and replaced them with FetchContent

+++ src/Nexus/src/Platform/D3D12/PipelineD3D12.cpp
@@ -272,7 +272,7 @@ namespace Nexus::Graphics
 												 elementIndex,
 												 D3D12::GetD3D12BaseType(element),
 												 layoutIndex,
-												 element.Offset,
+												 (UINT)element.Offset,
 												 classification,
 												 layout.GetInstanceStepRate()};
